fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
databases
asyncpg
